Scene erstellt über Random Functions ein Mesh
One Half of the Creature is created via SymmetricCreature class and then Displayed with the CreatureGeneration Class.

@daktylus11 Extensive Experimentation required:
Values to experiment on:

Random Funktions for x,y,z koordiantes

Box size and shape

Density of the Balls

Metaballsize

It is not necessary to change the SymmetricCreature class to change the Values.

However another point of experimetation is the value Function in balls.